TabularData now a separate file
new function to obtain insert_safe list of records where the parent is always  before the child in the list  to avoid errors during inserts.
